refactor: redesign layout to improve UI/UX
The commit updates project images in the Projects section with more specific, context-based, and high-quality images. This will improve user engagement by visually enhancing the UX.
The Home and Projects sections have also been restructured and restyled for a more intuitive and responsive web layout. SVGs no longer used in the project are now deleted. This helps in reducing overall bundle size and improving the app's performance.
On the other hand, a new SVG logo has been added as part of a rebranding strategy to enhance user identification of the brand.