feat: add navigation bar component
Implemented a new Navigation bar in the application to improve the user interface, which includes links to 'home', 'projects', and 'contact'. In addition, significant reorganization of the '/app' directory was carried out to improve code maintainability. This included renaming and relocating component files and modifying asset paths within the 'app' directory. This will make the project easier to work with as it continues to grow.